refactor(session-live): #2430 Block B+ T3 hoist useDebouncedCallback
Move the inline helper out of scores/page.tsx into a shared lib module
under lib/session-live/. Extend with a flush() method exposed via
[fn, flush] tuple return so ScoreTabContent (T7) can invoke pending
callbacks during unmount cleanup (DEC-4 flush-on-unmount).

scores/page.tsx callsite migrated to tuple destructure (flush ignored
at this callsite — only the new ScoreTabContent uses it).

B1 verification: grep -c 'function useDebouncedCallback' on the page
returns 0 (inline helper fully removed). Atomic edit applied.
